src/dashboard/__init__.py

from flask import Flask
from flask_socketio import SocketIO
from dotenv import load_dotenv
import os

load_dotenv()

socketio = SocketIO(cors_allowed_origins="*")

def create_app():
    app = Flask(__name__)
    app.config['SECRET_KEY'] = os.getenv('SECRET_KEY', 'dev-key-change-in-production')
    
    socketio.init_app(app, async_mode='eventlet')
    
    from .views import main_bp
    app.register_blueprint(main_bp)
    
    return app
