requirements.txt

Flask>=2.3.0
Flask-SocketIO>=5.3.0
paho-mqtt>=1.6.0
redis>=4.3.0
celery>=5.3.0
pandas>=2.0.0
pytest>=7.4.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
eventlet>=0.33.0
flask-cors>=4.0.0
